microservice-python/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Database - Compatible versions (fixed pymongo/bson conflict)
motor==3.3.1
pymongo==4.5.0

# Authentication and Security
bcrypt==4.0.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and Monitoring
structlog==23.2.0

# Testing (optional for development)
pytest==7.4.3
pytest-asyncio==0.21.1

# Data validation and serialization
email-validator==2.1.0

# Background tasks and scheduling (for voucher service)
celery==5.3.4
redis==5.0.1

# RabbitMQ for event-driven architecture
aio-pika==9.4.0

# Image processing (if needed for voucher images)
Pillow==10.1.0

# Date and time utilities
python-dateutil==2.8.2

# JSON processing
orjson==3.9.10

# Development dependencies (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0 